5.1.2 refactor: move render function, rename utils.js, add remove function
