+++ src/mock/comment.js
@@ -1,5 +1,5 @@
 import {EMOJIS} from '../const.js';
-import {getRandomArrayItem} from '../utils/utils.js';
+import {getRandomArrayItem} from '../utils/common.js';
 import {getRandomDate} from '../utils/film-utils.js';
 
 const text = [`Interesting setting and a good cast`, `Booooooooooring`, `Very very old. Meh`, `Almost two hours? Seriously?`];
